+++ examples/basic-project/src/app.tsx
@@ -11,6 +11,11 @@ console.error('__ERROR__');
 console.log('process.env.HAHA', process.env.HAHA);
 console.log('process.env.undefinedEnv', process.env.undefinedEnv);
 
+import('./standard-module')
+  .then((mod) => {
+    mod.printOne();
+  });
+
 if (isWeb) {
   console.error('__IS_WEB__');
 }