+++ examples/basic-project/src/standard-module.ts
@@ -1,0 +1,3 @@
+export function printOne() {
+  console.log(1);
+}