+++ packages/webpack-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.6
+
+- [fix] dynamic import works in app.tsx
+
 ## 1.0.5
 
 - [fix] bump version of `@ice/bundles`(0.1.2 -> 0.1.3)